+++ src/main/java/ru/kata/spring/boot_security/demo/services/UserService.java
@@ -1,10 +1,9 @@
 package ru.kata.spring.boot_security.demo.services;
 
-import org.springframework.security.core.userdetails.UserDetailsService;
 import ru.kata.spring.boot_security.demo.models.User;
 import java.util.List;
 
-public interface UserService extends UserDetailsService {
+public interface UserService {
 
     List<User> getAllUsers();
