+++ src/main/java/ru/kata/spring/boot_security/demo/services/UsersDetailsService.java
@@ -1,0 +1,30 @@
+package ru.kata.spring.boot_security.demo.services;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import ru.kata.spring.boot_security.demo.models.User;
+
+@Service
+public class UsersDetailsService implements UserDetailsService {
+
+    private final UserService userService;
+
+    @Autowired
+    public UsersDetailsService(UserService userService) {
+        this.userService = userService;
+    }
+
+    @Override
+    @Transactional
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        User user = userService.findUserByUsername(username);
+        if (user == null) {
+            throw new UsernameNotFoundException(String.format("User '%s' not found!", username));
+        }
+        return user;
+    }
+}
